move TrueCopies implementation into its own file
